[#151]Refactor(badge): 작성한 게시글 수 int 값을 long으로 받던 코드 int로 받도록 수정

+++ src/main/java/today/sesac/versebyverse/badge/service/BadgeService.java
@@ -52,7 +52,7 @@ public class BadgeService {
         boolean hasTargetBadge = memberBadgeList.stream()
                 .anyMatch(memberBadge -> memberBadge.getBadge().getName().equals(targetBadgeName));
 
-        long postCount = postRepository.countByAuthorIdAndIsDeletedFalse(authorId);
+        int postCount = postRepository.countByAuthorIdAndIsDeletedFalse(authorId);
 
         if (postCount > 0 && !hasTargetBadge) {
             Badge badge = badgeRepository.findByName(targetBadgeName).orElseThrow(
@@ -72,7 +72,7 @@ public class BadgeService {
         boolean hasTargetBadge = memberBadgeList.stream()
                 .anyMatch(memberBadge -> memberBadge.getBadge().getName().equals(targetBadgeName));
 
-        long postCount = postRepository.countByAuthorIdAndIsDeletedFalse(authorId);
+        int postCount = postRepository.countByAuthorIdAndIsDeletedFalse(authorId);
 
         if (postCount >= 10 && !hasTargetBadge) {
             Badge badge = badgeRepository.findByName(targetBadgeName).orElseThrow(
